Organize test cases according to the purpose of each test case (#175)
