+++ test/unit/util/llrb_tree_test.ts
@@ -15,7 +15,7 @@
  */
 
 import { assert } from 'chai';
-import { LLRBTree } from '../../src/util/llrb_tree';
+import { LLRBTree } from '../../../src/util/llrb_tree';
 
 const arrays = [
   [0, 1, 2, 3, 4, 5, 6, 7, 8, 9],
